+++ src/Components/AvailableUser/styles.js
@@ -1,0 +1,50 @@
+const styles = {
+  selectField: {
+    fontSize: '1.2rem',
+    width: '10vw',
+    height: '5vh',
+    backgroundColor: '#ffffff',
+  },
+  iconStyles: {
+    fill: '#00000',
+    left: '3vw',
+  },
+  labelStyles: {
+    textAlign: 'center',
+    paddingRight: '0',
+    color: '#49abb0',
+    fontFamily: 'DIN Pro',
+    fontWeight: 'normal',
+    fontStyle: 'normal',
+    fontStretch: 'normal',
+    letterSpacing: '0.3px',
+  },
+  hintStyles: {
+    backgroundColor: '#F1F6F8',
+    color: '#000',
+    fontSize: '.9rem',
+    height: '40px',
+    width: '90px',
+    lineHeight: '40px',
+    paddingTop: '0',
+    paddingRight: '0',
+    top: '3px',
+    textAlign: 'center',
+    verticalAlign: 'none',
+    borderRadius: '6px',
+  },
+  menuStyles: {
+    width: '7vw',
+    left: '2vw',
+  },
+  menuItemStyles: {
+    fontSize: '1rem',
+    fontFamily: 'DIN Pro',
+    fontWeight: '300',
+    fontStyle: 'normal',
+    fontStretch: 'normal',
+    letterSpacing: '0.3px',
+  },
+};
+
+export default styles;
